+++ CHANGELOG.md
@@ -6,6 +6,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
 
+### Added
+
+- Added `create-account` alias to `create-user` command.
+- Added `list-accounts` alias to `list-users` command.
+- Added `delete-user` alias to `delete-profile` command.
+
 ### Fixed
 
 - Canonicalize IDs when parsing `inReplyTo` and object links.